Add build_load_dag_redshift
